style(lint): clang-tidy autofix — redundant-string-init + braces + move-const (79 findings)
Applied clang-tidy --fix for the mechanical curated checks (readability-redundant-string-init,
readability-braces-around-statements, performance-move-const-arg, readability-redundant-casting)
across the tree. Cleared 79 of ~94 LINT findings introduced by the H/I RELAY work — the LINT
gate wasn't exercised while cpp CI was broken on the missing scheduler. Builds clean (g++) +
all 1863 tests pass. 13 findings needing judgment (optional-access guards, string-concat,
exception-escape, a false-positive macro-parens) handled in the follow-up commit.

+++ src/core/config_loader.cpp
@@ -140,7 +140,7 @@ json ConfigLoader::substitute_string(const std::string& value) const {
   }
   if (is_all_digits(result)) {
     try {
-      return json(static_cast<long long>(std::stoll(result)));
+      return json(std::stoll(result));
     } catch (const std::exception&) {
       return json(result);
     }
